Conflicts:
	common.mak
	libvo/vo_xv.c
	libvo/x11_common.c
	libvo/x11_common.h
	stream/cache2.c

+++ libmpdemux/demux_audio.c
@@ -605,7 +605,7 @@ static void high_res_mp3_seek(demuxer_t *demuxer,float time) {
 static void demux_audio_seek(demuxer_t *demuxer,float rel_seek_secs,float audio_delay,int flags){
   sh_audio_t* sh_audio;
   stream_t* s;
-  int base,pos;
+  int64_t base,pos;
   float len;
   da_priv_t* priv;
